remove fetcher and specify external package
